feat: separate default editor in other module
