+++ packages/creator-hub/main/src/index.ts
@@ -9,7 +9,7 @@ import { deployServer, killAllPreviews } from '/@/modules/cli';
 import { killInspectorServer } from '/@/modules/inspector';
 import { runMigrations } from '/@/modules/migrations';
 import { getAnalytics, track } from './modules/analytics';
-import { addEditorsPathsToConfig } from './modules/bin';
+import { addEditorsPathsToConfig } from './modules/default-editor-settings';
 
 import '/@/security-restrictions';
